requirements.txt
# Production & Development dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.30.0
requests==2.31.0
httpx==0.25.2
streamlit==1.44.1
